Traffic Collision Data from 2020 to Present
Description: This dataset is a subset of the dataset titled "Traffic Collision Data from 2010 to Present" by the Los Angeles Police Department. It includes traffic collision incidents in the City of Los Angeles between 2020 and 2023.
How it'll be used for UP 221 project: Dataset includes the coordinates of each collision. For my project idea to map intersections with crashes above a certain threshold, crash data can be grouped based on their coordinates and counted to determine the number of crashes at each intersection. 
Source: https://data.lacity.org/Public-Safety/Traffic-Collision-Data-from-2010-to-Present/d5tf-ez2w